Applicant Dashboard
Applicants should have access to a dashboard or notice board section.

Rq ID : TSS_009.1

+++ back-end/tss/src/main/java/com/bjit/tss/service/impl/ApplicationServiceImpl.java
@@ -70,7 +70,7 @@ public class ApplicationServiceImpl implements ApplicationService {
         }
 
         ListResponse<?> listResponse = ListResponse.builder()
-                .dataLength((long) examineeInfos.get().size())
+                .dataLength(examineeInfos.get().size())
                 .listResponse(examineeInfos)
                 .build();
         return ApiResponseMapper.mapToResponseEntityOK(listResponse);
